+++ HISTORY.md
@@ -7,6 +7,7 @@
 #### API
 
 - Add UnstructuredArchive for novelty search ({pr}`472`)
+- Support diversity optimization in Scheduler.tell ({pr}`473`)
 - Allow specifying separate dtypes for solution, objective, and measures
   ({pr}`471`)
 - Replace archive.dtype with archive.dtypes dict that holds dtype of every field